refactor(mysql): implement QueryEngine on MysqlEngine
MySQL's engine used to return JSON blobs with no path to typed
models. Rewrites it on top of MysqlRowRef / FromRow so
client.user().find_many::<User>() decodes rows end-to-end. The
legacy JSON-returning surface moves to prax_mysql::raw
(MysqlRawEngine / MysqlJsonRow) as an escape hatch for callers
that genuinely want untyped rows.

Keeps short-circuited query_one / query_optional / execute_insert
paths (inspired by the SQLite first-row helper in 1c202bd) so
queries accidentally passed to single-row methods don't
materialize the tail.
